added id to the user session and moved the auth option to a different file
