+++ task2/src/types/next-auth.d.ts
@@ -1,0 +1,9 @@
+import { DefaultSession } from "next-auth";
+
+declare module "next-auth" {
+   interface Session {
+      user: {
+         id: string;
+      } & DefaultSession["user"];
+   }
+}
